Move utxo collateral check into BlockDiskAccessor
