+++ divi/src/MasternodeHelpers.cpp
@@ -6,36 +6,13 @@
 #include <masternode.h>
 #include <blockmap.h>
 #include <BlockDiskAccessor.h>
-#include <coins.h>
+#include <MasternodePing.h>
 
 extern CCriticalSection cs_main;
 extern bool fImporting;
 extern bool fReindex;
 extern CChain chainActive;
 extern BlockMap mapBlockIndex;
-extern CCoinsViewCache* pcoinsTip;
-
-bool CollateralIsExpectedAmount(const COutPoint &outpoint, int64_t expectedAmount)
-{
-    CCoins coins;
-    LOCK(cs_main);
-    if (!pcoinsTip->GetCoins(outpoint.hash, coins))
-        return false;
-
-    int n = outpoint.n;
-    if (n < 0 || (unsigned int)n >= coins.vout.size() || coins.vout[n].IsNull()) {
-        return false;
-    }
-    else if (coins.vout[n].nValue != expectedAmount)
-    {
-        return false;
-    }
-    else {
-        return true;
-    }
-    assert(false);
-}
-
 
 bool IsBlockchainSynced()
 {
